# survey-parent/api-survey/src/main/resources/static/json/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "survey",
    "version": "v1.0",
    "title": "API文档",
    "termsOfService": "lego.cn",
    "contact": {
      "name": "lego",
      "url": "lego.cn",
      "email": "survey@legocloud.cn"
    }
  },
  "host": "192.168.101.197:50010",
  "basePath": "/",
...
  "paths": {
...
    "/survey/api-settlement/surveyOriginal/create/{sectionId}": {
      "post": {
        "tags": [
          "survey-original-controller"
        ],
        "summary": "添加原始数据",
        "description": "添加原始数据",
        "operationId": "addUsingPOST_5",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "sectionId",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "surveyOriginalVo",
            "description": "surveyOriginalVo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SurveyOriginalVo"
            }
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-settlement/surveyOriginal/delete/{sectionId}": {
      "delete": {
        "tags": [
          "survey-original-controller"
        ],
        "summary": "删除原始数据",
        "description": "删除原始数据",
        "operationId": "deleteUsingDELETE_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "ids",
            "in": "query",
            "description": "原始数据id",
            "required": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "标段ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-settlement/surveyOriginal/list/{sectionId}/{pageIndex}": {
      "get": {
        "tags": [
          "survey-original-controller"
        ],
        "summary": "列举所有原始数据",
        "description": "列举所有原始数据",
        "operationId": "listUsingGET_6",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "pageIndex",
            "in": "path",
            "description": "当前页",
            "required": true,
            "type": "integer",
            "format": "int32",
            "x-example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "type": "integer",
            "default": 10,
            "format": "int32",
            "x-example": 10
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "标段id",
            "required": false,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO«RespDataVO«SurveyOriginalVo»»"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
    "/survey/api-settlement/surveyOriginal/modify/{sectionId}": {
      "put": {
        "tags": [
          "survey-original-controller"
        ],
        "summary": "修改原始数据",
        "description": "修改原始数据",
        "operationId": "modifyUsingPUT_5",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "deviceType",
            "in": "header",
            "description": "设备类型(1-Android;2-WEB)",
            "required": true,
            "type": "string",
            "maxLength": 2,
            "minLength": 1
          },
          {
            "name": "sectionId",
            "in": "path",
            "description": "sectionId",
            "required": true,
            "type": "string"
          },
          {
            "name": "sn",
            "in": "header",
            "description": "设备序列号",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "surveyOriginalVo",
            "description": "surveyOriginalVo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SurveyOriginalVo"
            }
          },
          {
            "name": "timeStamp",
            "in": "header",
            "description": "当前时间戳",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "header",
            "description": "令牌",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "1": {
            "description": "操作成功"
          },
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespVO"
            }
          },
          "404": {
            "description": "Resources Not Found",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          },
          "-1": {
            "description": "权限校验失败"
          },
          "-2": {
            "description": "登录错误"
          },
          "-3": {
            "description": "服务内部错误"
          },
          "-4": {
            "description": "调用超时错误"
          },
          "-5": {
            "description": "其他错误"
          },
          "-6": {
            "description": "结果错误"
          },
          "-7": {
            "description": "数据处理异常"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "ApiError": {
      "type": "object",
      "title": "ApiError"
    },
...
    "OriginalDataVo": {
      "type": "object",
      "properties": {
        "distance": {
          "type": "string"
        },
        "index": {
          "type": "integer",
          "format": "int32",
          "description": "索引"
        },
        "surveyTime": {
          "type": "string",
          "format": "date-time",
          "description": "测量时间"
        },
        "tag": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "description": "原始数据value"
        }
      },
      "title": "OriginalDataVo"
    },
...
    "RespDataVO«SurveyOriginalVo»": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SurveyOriginalVo"
          }
        }
      },
      "title": "RespDataVO«SurveyOriginalVo»"
    },
...
    "RespVO": {
      "type": "object",
      "properties": {
        "info": {
          "type": "object"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO"
    },
...
    "RespVO«RespDataVO«SurveyOriginalVo»»": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/RespDataVO«SurveyOriginalVo»"
        },
        "msg": {
          "type": "string"
        },
        "ret": {
          "type": "string"
        },
        "retCode": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "RespVO«RespDataVO«SurveyOriginalVo»»"
    },
...
    "SurveyOriginalVo": {
      "type": "object",
      "properties": {
        "bpCode": {
          "type": "string",
          "description": "后视点编码"
        },
        "bpType": {
          "type": "integer",
          "format": "int32",
          "description": "后视点类型 0 基准点 1 观测点  2 转点"
        },
        "datas": {
          "type": "array",
          "description": "原始测量数据 JSON数组",
          "items": {
            "$ref": "#/definitions/OriginalDataVo"
          }
        },
        "fpCode": {
          "type": "string",
          "description": "前视点编码"
        },
        "fpType": {
          "type": "integer",
          "format": "int32",
          "description": "前视点类型 0 基准点 1 观测点  2 转点"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "原始数据id"
        },
        "index": {
          "type": "integer",
          "format": "int32",
          "description": "任务中索引值"
        },
        "taskId": {
          "type": "integer",
          "format": "int64",
          "description": "原始数据任务id"
        }
      },
      "title": "SurveyOriginalVo"
    },
...
  }
}